Example algorithm/requirements.txt
pillow
SimpleITK
imagecodecs
numpy